requirements.txt
langchain
langchain-core
langchain-openai
langchain-google-genai

# --- Environment & Config ---
python-dotenv
pydantic 

# --- LangGraph Agent ---
langgraph
langchain-postgres
langchain_community
langgraph-checkpoint-postgres
psycopg[binary]
langgraph-cli[inmem]
langsmith

# --- Other ---
tavily-python
requests
langchain-tavily